+++ docs/CHANGELOG.md
@@ -10,6 +10,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 ### Changed
 ### Fixed
 - [NO-ISSUE] - Fix a bug where we didn't respect the overrun.
+- [Issue 279](https://github.com/aza547/wow-recorder/issues/279) - Improve signalling robustness.
 
 ## [3.3.2] - 2023-01-08
 ### Fixed